+++ OpenSource/Assets/Scripts/mob/Grunt.cs
@@ -8,8 +8,8 @@ public class Grunt : MonoBehaviour
     public GameObject coin;
     public GameObject getHitEffect;
     public BoxCollider attackRange;
-    public int maxHp;
-    public int nowHp;
+    public float maxHp;
+    public float nowHp;
 
     Transform target;
     Vector3 des;
